outqueue: use u32 for len in quic_outq_limit_check()
Change len from u16 to u32 to avoid possible truncation when summing packet
and frame lengths for congestion control and anti-amplification checks.

+++ modules/net/quic/outqueue.c
@@ -18,7 +18,7 @@ static int quic_outq_limit_check(struct sock *sk, struct quic_frame *frame)
 	struct quic_path_group *paths = quic_paths(sk);
 	struct quic_packet *packet = quic_packet(sk);
 	struct quic_outqueue *outq = quic_outq(sk);
-	u16 len;
+	u32 len;
 
 	/* If in single-packet mode, allow only one packet to transmit. */
 	if (outq->single && outq->count)
